requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.0
qdrant-client==1.11.0
httpx==0.27.0
python-dotenv==1.0.1

